add trimesh.iteration
