+++ trimesh/path/polygons.py
@@ -3,11 +3,11 @@ from shapely import ops
 from shapely.geometry import Polygon
 
 from .. import bounds, geometry, graph, grouping
-from ..util import reduce_cascade
 from ..constants import log
 from ..constants import tol_path as tol
 from ..transformations import transform_points
 from ..typed import Iterable, NDArray, Number, Optional, Union, float64, int64
+from ..util import reduce_cascade
 from .simplify import fit_circle_check
 from .traversal import resample_path
